refactor(add-chain): refactor default command (#717)
* make .env file not required

* move addresses into util

* extract default command

* make SuperchainTargetFlag and ChainShortNameFlag required

* fix lint errors

* make default command normal command and rename to add-new-chain

* update superchainRepoRoot path detection

* lint
