+++ add-chain/cmd/compress-genesis.go
@@ -39,13 +39,7 @@ var CompressGenesisCmd = cli.Command{
 		}
 		superchainRepoRoot := filepath.Dir(filepath.Dir(filepath.Dir(thisFile)))
 		superchainTarget := ctx.String(flags.SuperchainTargetFlag.Name)
-		if superchainTarget == "" {
-			return fmt.Errorf("missing required flag: %s", flags.SuperchainTargetFlag.Name)
-		}
 		chainShortName := ctx.String(flags.ChainShortNameFlag.Name)
-		if chainShortName == "" {
-			return fmt.Errorf("missing required flag: %s", flags.ChainShortNameFlag.Name)
-		}
 
 		zipOutputDir := filepath.Join(superchainRepoRoot, "/superchain/extra/genesis", superchainTarget, chainShortName+".json.gz")
 		genesisPath := ctx.Path(flags.GenesisFlag.Name)
